+++ R/pandoc.R
@@ -19,20 +19,23 @@
 #' If there are multiple records of the \code{t} field in the configuration, the
 #' input markdown file will be converted to all these formats by default, unless
 #' the \code{format} argument is specified as one single format.
-#' @param input A character vector of Markdown filenames.
-#' @param format Name of the output format (see References). This can be a character
-#'   vector of multiple formats; by default, it is obtained from the \code{t}
-#'   field in the configuration. If the configuration is empty or the \code{t}
-#'   field is not found, the default output format will be \code{'html'}.
-#' @param config Path to the Pandoc configuration file. If missing, it is assumed to be
-#'   a file with the same base name as the \code{input} file and an extension
-#'   \code{.pandoc} (e.g. for \file{foo.md} it looks for \file{foo.pandoc})
-#' @param ext Filename extensions. By default, the extension is inferred
-#'   from the \code{format}, e.g. \code{latex} creates \code{pdf},
-#'   \code{dzslides} creates \code{html}, and so on
+#' @param input A character vector of Markdown filenames (must be encoded in
+#'   UTF-8).
+#' @param format Name of the output format (see References). This can be a
+#'   character vector of multiple formats; by default, it is obtained from the
+#'   \code{t} field in the configuration. If the configuration is empty or the
+#'   \code{t} field is not found, the default output format will be
+#'   \code{'html'}.
+#' @param config Path to the Pandoc configuration file. If missing, it is
+#'   assumed to be a file with the same base name as the \code{input} file and
+#'   an extension \code{.pandoc} (e.g. for \file{foo.md} it looks for
+#'   \file{foo.pandoc})
+#' @param ext Filename extensions. By default, the extension is inferred from
+#'   the \code{format}, e.g. \code{latex} creates \code{pdf}, \code{dzslides}
+#'   creates \code{html}, and so on
 #' @inheritParams knit
 #' @return The output filename(s) (or an error if the conversion failed).
-#' @references Pandoc: \url{http://pandoc.org}; Examples and rules of the
+#' @references Pandoc: \url{https://pandoc.org}; Examples and rules of the
 #'   configurations: \url{https://yihui.name/knitr/demo/pandoc/}
 #'
 #'   Also see R Markdown (v2) at \url{http://rmarkdown.rstudio.com}. The
@@ -43,15 +46,13 @@
 #' @seealso \code{\link{read.dcf}}
 #' @export
 #' @examples system('pandoc -h') # see possible output formats
-pandoc = function(input, format, config = getOption('config.pandoc'), ext = NA,
-                  encoding = getOption('encoding')) {
+pandoc = function(input, format, config = getOption('config.pandoc'), ext = NA) {
   if (Sys.which('pandoc') == '')
     stop('Please install pandoc first: http://pandoc.org')
   cfg = if (is.null(config)) with_ext(input[1L], 'pandoc') else config
-  con = file(input[1L], encoding = encoding)
-  tryCatch(txt <- pandoc_cfg(readLines(con, warn = FALSE)), finally = close(con))
-  if (file.exists(cfg)) txt = c(txt, '', readLines(cfg, warn = FALSE))
-  con = textConnection(txt); on.exit(close(con))
+  txt = xfun::read_utf8(input[1])
+  if (file.exists(cfg)) txt = c(txt, '', xfun::read_utf8(cfg))
+  con = textConnection(txt, encoding = 'UTF-8'); on.exit(close(con), add = TRUE)
   cfg = read.dcf(con)
   nms = colnames(cfg)
   if (length(nms) && 'format' %in% nms) {
@@ -59,21 +60,13 @@ pandoc = function(input, format, config = getOption('config.pandoc'), ext = NA,
     colnames(cfg)[nms == 'format'] = 't'  # for backward compatibility
   }
   if (missing(format)) format = pandoc_fmt(cfg)
-  input_utf8 = input
-  if (encoding != 'UTF-8') {
-    for (i in seq_along(input)) {
-      input_utf8[i] = gsub('[.]([[:alnum:]]+)$', '_utf8.\\1', input[i])
-      encode_utf8(input[i], encoding, input_utf8[i])
-    }
-    on.exit(unlink(input_utf8), add = TRUE)
-  }
   mapply(
-    pandoc_one, input, input_utf8, format, ext, MoreArgs = list(cfg = cfg),
+    pandoc_one, input, format, ext, MoreArgs = list(cfg = cfg),
     USE.NAMES = FALSE
   )
 }
 # format is a scalar
-pandoc_one = function(input, input_utf8, format, ext, cfg) {
+pandoc_one = function(input, format, ext, cfg) {
   cmn = NULL  # common arguments
   if (nrow(cfg) == 0L) cfg = character(0) else if (nrow(cfg) == 1L) {
     if ('t' %in% colnames(cfg)) {
@@ -98,9 +91,11 @@ pandoc_one = function(input, input_utf8, format, ext, cfg) {
     }
   })
   cfg = cfg[setdiff(names(cfg), c('o', 'output', 't'))]
-  cmd = paste('pandoc', pandoc_arg(cfg), pandoc_arg(cmn),
-              '-t', format, '-o', out, paste(shQuote(input_utf8), collapse = ' '))
-  message('executing ', cmd)
+  cmd = paste(
+    'pandoc', pandoc_arg(cfg), pandoc_arg(cmn), '-t', format, '-o', out,
+    paste(shQuote(input), collapse = ' ')
+  )
+  message('Executing ', cmd)
   if (system(cmd) == 0L) out else stop('conversion failed')
 }